requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
jsonify==0.5
MarkupSafe==2.1.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-dotenv==0.19.2
requests==2.27.1
SQLAlchemy==1.4.32
toml==0.10.2
urllib3==1.26.8
Werkzeug==2.0.3
